agent/pyproject.toml
[project]
name = "wakeword-agent"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "livekit-agents[silero,turn-detector]~=1.3",
    "python-dotenv~=1.0",
]

[project.scripts]
wakeword-agent = "agent:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel.force-include]
"agent.py" = "agent.py"
